requirements.txt
asgiref==3.7.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.2
djangorestframework==3.15.1
drf-yasg==1.21.7
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
libretranslatepy==2.1.1
lxml==5.2.1
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.4.4
translate==3.6.1
uritemplate==4.1.1
urllib3==2.2.1
